--- a/src/base/thread_pool.h
+++ b/src/base/thread_pool.h
@@ -76,6 +76,8 @@ private:
     // synchronization
     std::mutex queue_mutex;
     std::condition_variable condition;
+    std::mutex sync_mutex;
+    std::condition_variable sync_condition;
     bool stop;
     std::atomic_int sync { 0 };
 };
@@ -101,6 +103,10 @@ inline ThreadPool::ThreadPool(size_t threads)
             }
          task();
          sync++;
+         {
+           std::unique_lock<std::mutex> lock(this->sync_mutex);
+           sync_condition.notify_one();
+         }
       }
     }
   );
@@ -129,7 +135,10 @@ auto ThreadPool::enqueue(F&& f, Args&&... args)
 
 // Wait all thread to finish their jobs
 inline void ThreadPool::Sync(int wait_count) {
-  while (sync != wait_count) {}
+  std::unique_lock<std::mutex> lock(sync_mutex);
+  this->sync_condition.wait(lock, [&]() {
+    return sync == wait_count;
+  });
   sync = 0;
 }