accelerate threadpool synchronization