@aksnzhy 2018-05-08
@purkyston Can you run a simple benchmark to test the performance improvement and put the result on this pr?  Thanks!