@purkyston 2018-05-10
| time cost of each epoch (sec)  | before | after | speedup |
| --- | --- | --- | --- |
| LR | 0.1365 | 0.1281 | 6.2% |
| FM | 0.2832 | 0.2486 | 12.2% |
| FFM | 1.521 | 1.411 | 7.2% |

Hardware: 4 cores (one thread per core)
Threadpool size: 4
speedup = (before - after) / before